Linting

+++ Gmock/lib/notifier.cpp
@@ -3,16 +3,13 @@
 #include "lib/notifier.h"
 
 Notifier::Notifier(std::ostream & os) :
-    os_{os}
-{
+    os_{ os } {
 }
 
-void Notifier::deposit_made(unsigned long amount)
-{
+void Notifier::deposit_made(unsigned int amount) {
     os_ << "Deposited: " << amount << std::endl;
 }
 
-void Notifier::withdrawal_made(unsigned long amount)
-{
+void Notifier::withdrawal_made(unsigned int amount) {
     os_ << "Withdrawn: " << amount << std::endl;
 }
