+++ Gmock/lib/notifier_interface.h
@@ -4,10 +4,10 @@
 
 class Notifier_Interface
 {
-protected:
+ protected:
     virtual ~Notifier_Interface() = default;
 
-public:
-    virtual void deposit_made(unsigned long amount) = 0;
-    virtual void withdrawal_made(unsigned long amount) = 0;
+ public:
+    virtual void deposit_made(unsigned int amount) = 0;
+    virtual void withdrawal_made(unsigned int amount) = 0;
 };
